Backend/requirements.txt
requests
openai
python-dotenv
ipykernel
Flask
google-generativeai